paicc-2-copy/pyproject.toml
[project]
name = "transcript_analytics"
version = "0.1.0"
description = "Simple transcript analytics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai==1.55.3",
    "pydantic==2.9.2",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/transcript_analytics"]

[project.scripts]
main = "transcript_analytics.main:main"
